+++ nuxt.config.ts
@@ -54,9 +54,6 @@ export default defineNuxtConfig({
   ssr: !isDev,
   app: {
     head: {
-      htmlAttrs: {
-        lang: i18nLocales.find(item => item.code === config.defaultLang)?.iso
-      },
       meta: [
         { charset: "utf-8" },
         {
@@ -64,7 +61,8 @@ export default defineNuxtConfig({
           content: "width=device-width,initial-scale=1,minimum-scale=1,maximum-scale=1,user-scalable=no"
         },
         { name: "description", content: config.SEO_description },
-        { name: "keywords", content: config.SEO_keywords }
+        { name: "keywords", content: config.SEO_keywords },
+        { name: "author", content: config.nickName }
       ],
       link: [
         { rel: "shortcut icon", href: "/icon.png" }